init project jumpo
Jump list for moving between directories in very quick and easy way.
Depends on the prefix we set for the location. This will enable less
keystrokes to move into any directory and much easier since fuzzy
matching will take some time but here there is direct lookup so it is
pretty fast.

Currently four functions are implemented:

`-add` : Adds prefix to the location where we are executing the command
`-remove` : Remove prefix from the cache
`-list` : List all the prefix that are stored in the cache
`prefix` : Here prefix is the key that is present in the cache where you
want to jump. This will lead you to that directory.

Since executables are running in a separate process (forked from the
parent shell) we need to use a bash function so that we can change the
current directory of the parent shell.